Lift measurement from PageNode statics to extension state
